issolate popper logic in a component
