+++ src/view/util/Popper.scss
@@ -1,18 +1,3 @@
-.caseCreationForm {
-  text-align: center;
-  padding: 10px;
-
-  .usa-button {
-    width: 150px;
-    margin-top: 10px;
-    margin-right: 0px;
-
-    &:first-child {
-      margin-top: 0px;
-    }
-  }
-}
-
 .popper {
   margin-top: 20px;
   position: absolute;
@@ -50,7 +35,6 @@
   left: calc(50% - 10px);
   top: 0;
 }
-
 
 // .popper .popper__arrow {
 //   border-color: #fff;
